Changed the Makefile so -Wnewline-eof was conditional on the Operating System being Darwin Modified the variable name from file_size to print_file_size so a cast wasn't required over different arch's Casted size_t to long int in gerber_parse.cpp Removed the inclusion of gd.h in gerb_script_util.cpp as gd is no longer required

+++ src/gerber_parse.cpp
@@ -845,7 +845,7 @@ bool parse_274X_param_do_block(char ** cur_block_ptr, RS274X_Program * target, i
 			break;	
 
 		default:
-			DBG_ERR_PF("unmatched 274X paramblock %s - len %ld", *cur_block_ptr, strlen(*cur_block_ptr));
+			DBG_ERR_PF("unmatched 274X paramblock %s - len %ld", *cur_block_ptr, (long int)strlen(*cur_block_ptr));
 			//parse_ok = false;
 	}
